Add tests for utils
